- move the javafx mvc packages to a common components package
