+++ phoenicis-javafx/src/main/java/org/phoenicis/javafx/components/skin/SkinBase.java
@@ -1,6 +1,6 @@
-package org.phoenicis.javafx.skin;
+package org.phoenicis.javafx.components.skin;
 
-import org.phoenicis.javafx.controler.ControlBase;
+import org.phoenicis.javafx.components.control.ControlBase;
 
 /**
  * A base class for a JavaFX component skin/view
